Refactor core fingerprint relationships
